Actualización y reorganización de servicios y controladores: - Modificado 'app.js' para integrar nuevas rutas de materiales. - Actualizado 'controladorCursos.js' para nuevas funcionalidades. - Reemplazado 'servicioCursos.js' y añadido 'servicioMateriales.js' para manejo de materiales. - Añadido controlador y rutas para materiales ('controladorMateriales.js' y 'materialesRutas.js').
